Add files via upload
Chinmaya Mission Hospitals are often referred to as the best hospitals for the poorest and middle-class householders. With a new set of funds coming in, the CEO of Chinmaya mission Hospital has decided to open new branches in various parts of Karnataka. However, he doesn't have any data to target places that could be most beneficial to people. He has reached out to us to design an application that can record patients' locations. He wants to understand, on a given day or for a range of days, how many OP's are within Bangalore and how many are from outside. Based on this, he will make a decision if it is feasible to open new branches outside of Bangalore or open new Centers within Bangalore itself?